% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\name{as_lkthr}
\alias{as_lkthr}
\alias{as_lkthr.data.frame}
\alias{as_lkthr.list}
\title{Make a lkthr(abbr of lookthrough) object}
\usage{
as_lkthr(x, ...)

\method{as_lkthr}{data.frame}(x, mapping = c(ptf = "PTF", asset = "ASSET",
  exposure = "EXPOSURE"), ...)

\method{as_lkthr}{list}(x, ...)
}
\arguments{
\item{x}{A \code{list} or a \code{data.frame} contains position info. If
\code{x} is a \code{list}, it should contain three levels of each represents
the \code{PTF}, \code{ASSET} and \code{EXPOSURE}.}

\item{...}{Other arguments being passed to the patched functions. Currently,
if \code{x} is a \code{data.frame}, you can use a named character vector \code{mapping}
to map the \code{data.frame} columns \code{PTF}, \code{ASSET} and \code{EXPOSURE}.}
}
\value{
A \code{lkthr} object. Essentially it's a \link[data.tree:Node]{data.tree::Node}.
}
\description{
Make a lkthr(abbr of lookthrough) object
}
