% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\name{lkthr_filter}
\alias{lkthr_filter}
\title{Filter on condition}
\usage{
lkthr_filter(ptfs, fun)
}
\arguments{
\item{ptfs}{A \code{lkthr} object contains the ptf position.}

\item{fun}{A binary function with only one param \code{node}.}
}
\value{
A \code{lkthr} object.
}
\description{
Filter the \code{lkthr} tree by the attributes on leafs.
}
