% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\name{lkthr_match}
\alias{lkthr_match}
\title{Match the portfolio and the funds}
\usage{
lkthr_match(ptfs, funds, max_layer = 5L)
}
\arguments{
\item{ptfs}{A \code{lkthr} object contains the ptf position.}

\item{funds}{A \code{lkthr} object contains the fund position.}

\item{max_layer}{The maximum look-through layers that's allowed to perform.}
}
\value{
The altered \code{ptfs} invisibly.
}
\description{
The fund position will be appended the asset by compare the name.
}
\details{
It's a by reference operation and will alter the \code{ptfs} input. You
can call \code{\link[data.tree:Clone]{data.tree::Clone()}} in order not to change the original value.
}
