% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\docType{data}
\name{lkthr_sample}
\alias{lkthr_sample}
\title{lkthr sample data}
\format{An object of class \code{list} of length 3.}
\usage{
data(lkthr_sample)
}
\description{
A dataset contains artifical ptf positions, fund positions and
asset issuers / guarantors.
}
\author{
Xianying Tan
}
\keyword{datasets}
