% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\name{lkthr_set}
\alias{lkthr_set}
\title{Set the assets' attributes}
\usage{
lkthr_set(ptfs, attr)
}
\arguments{
\item{ptfs}{A \code{lkthr} object contains the ptf position.}

\item{attr}{A named \code{list}. The asset will be match by
its names.}
}
\value{
The altered \code{ptfs} invisibly.
}
\description{
Set character or numeric attributes to each asset. Again,
it's a by reference function.
}
