% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkthr.R
\name{as_lkthr}
\alias{as_lkthr}
\title{Make a lkthr(abbr of lookthrough) object}
\usage{
as_lkthr(x, ...)
}
\arguments{
\item{x}{A \code{list} or a \code{data.frame} contains position info. If
\code{x} is a \code{list}, it should contain three levels of each represents
the \code{PTF}, \code{ASSET} and \code{EXPOSURE}. If x is}

\item{mapping}{When x is a \code{data.frame}, you can}
}
\description{
Make a lkthr(abbr of lookthrough) object
}
